Altered the locale information and removed herndon and replaced with local.  Also included the other possible servers to the list (development and preprod).  These are primarily on the CCSA servers and can be ignored if moved. Also switch the default operator from LIKE to =.  More people requested it over Larry.
Also added the capability to store selected columns definition which will be automatically applied each time the user logs in until they clear it.  This is also true for the primary key.  For those table that do not have an existing primary key, the key is stored and will return each time they user logs on until it is cleared.